lightr: Read Spectrometric Data and Metadata

Parse various reflectance/transmittance/absorbance spectra file formats to extract spectral data and metadata, as described in Gruson, White & Maia (2019) doi:10.21105/joss.01857 . Among other formats, it can import files from 'Avantes' https://www.avantes.com/, 'CRAIC' https://www.microspectra.com/, and 'OceanOptics'/'OceanInsight' https://www.oceanoptics.com/ brands.

See also

Author

Maintainer: Hugo Gruson hugo.gruson+R@normalesup.org (ORCID) [copyright holder]

Authors:

Hugo Gruson hugo.gruson+R@normalesup.org (ORCID) [copyright holder]

Rafael Maia (ORCID) [copyright holder]

Thomas White (ORCID) [copyright holder]

Other contributors:

Kotya Karapetyan (Author of the MATLAB script to read AvaSoft7 binary files (CC-BY)) [contributor, copyright holder]